requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0       
charset-normalizer==3.3.2   
constantly==23.10.4
cryptography==42.0.8        
daphne==4.1.2
defusedxml==0.8.0rc2        
Django==5.0.6
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==70.2.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==6.4.post2


dj-database-url==2.2.0
django-cors-headers==4.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3